fix(repository): apply text/link/steps operator filters to exported cases
Exports fetched rows with the SQL where clause only, so the post-fetch
text/link/steps operator filters the table applies were silently ignored
and "all filtered" exports returned the unfiltered superset. The matchers
now live in a pure server-safe module (lib/repositoryCaseFieldMatchers.ts,
re-exported from the hook for existing consumers) and the export action
applies them before shared-step expansion so step counts match the table.
